Encrypt and decrypt files for AffineCipher, PermutationCipher and VigenereCipher algs

+++ src/model/AbstractEncryptionAlgorithm.java
@@ -53,7 +53,7 @@ public abstract class AbstractEncryptionAlgorithm implements EncryptionAlgorithm
 
 
     @Override
-    public String generateKey(int keyLength) {
+    public String generateKey(long keyLength) {
         throw new UnsupportedOperationException("Not implemented");
     }
